Fix listener registration on older servers
