+++ src/main/java/me/ryanhamshire/GriefPrevention/GriefPrevention.java
@@ -368,6 +368,12 @@ public class GriefPrevention extends JavaPlugin
         //player events
         playerEventHandler = new PlayerEventHandler(this.dataStore, this);
         pluginManager.registerEvents(playerEventHandler, this);
+        try {
+            Class.forName("org.bukkit.event.player.PlayerSignOpenEvent");
+            pluginManager.registerEvents(new PlayerSignOpenHandler(this), this);
+        } catch (ClassNotFoundException ignored) {
+            // Event is not present in this version, don't attempt to register events and cause a log warning.
+        }
 
         //block events
         BlockEventHandler blockEventHandler = new BlockEventHandler(this.dataStore);
