+++ src/main/java/me/ryanhamshire/GriefPrevention/PlayerEventHandler.java
@@ -80,7 +80,6 @@ import org.bukkit.event.player.PlayerLoginEvent.Result;
 import org.bukkit.event.player.PlayerPortalEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.player.PlayerRespawnEvent;
-import org.bukkit.event.player.PlayerSignOpenEvent;
 import org.bukkit.event.player.PlayerTakeLecternBookEvent;
 import org.bukkit.event.player.PlayerTeleportEvent;
 import org.bukkit.event.player.PlayerTeleportEvent.TeleportCause;
@@ -92,7 +91,6 @@ import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.util.BlockIterator;
-import org.jetbrains.annotations.NotNull;
 
 import java.net.InetAddress;
 import java.util.ArrayList;
@@ -1598,28 +1596,6 @@ class PlayerEventHandler implements Listener
             bucketEvent.setCancelled(true);
             return;
         }
-    }
-
-    @EventHandler(priority = EventPriority.LOW)
-    void onPlayerSignOpen(@NotNull PlayerSignOpenEvent event)
-    {
-        if (event.getCause() != PlayerSignOpenEvent.Cause.INTERACT || event.getSign().getBlock().getType() != event.getSign().getType())
-        {
-            // If the sign is not opened by interaction or the corresponding block is no longer a sign,
-            // it is either the initial sign placement or another plugin is at work. Do not interfere.
-            return;
-        }
-
-        Player player = event.getPlayer();
-        String denial = instance.allowBuild(player, event.getSign().getLocation(), event.getSign().getType());
-
-        // If user is allowed to build, do nothing.
-        if (denial == null)
-            return;
-
-        // If user is not allowed to build, prevent sign UI opening and send message.
-        GriefPrevention.sendMessage(player, TextMode.Err, denial);
-        event.setCancelled(true);
     }
 
     //when a player interacts with the world
